Fix tests to reflect changes in REPOSLUG parameter handling.
The asyncio version of the assignment requires the app to be capable of
handling multiple REPOSLUG parameters. The most straightforward way to
do so seems to be nargs=-1 parameter for the click.argument decorator,
which adds three dots after the parameter name in help message and all
error messages. This commit introduces the same change to the tests.